This commit message is indicating that the changes made in this commit are related to adding support for saving and reading metadata in SVG files. The diff output shows the addition of a new function `add_svgdata` which adds metadata to an SVG file, as well as a new function `read_svgdata` which reads metadata from an SVG file. Additionally, the commit message includes a reference to the `gzip` module, which is used for compressing and decompressing data in the SVG file.
The diff output also shows that the changes made in this commit are related to the `matplot2.py` script, specifically the addition of two new functions `add_svgdata` and `read_svgdata`. The `save`
function now supports saving metadata in SVG files when the `--write-data` argument is passed, and the `read_metadata` function has been updated to read metadata from SVG files.

Overall, this commit message indicates that a new feature has been added to the script to support reading and writing metadata in SVG files, and that the changes made in this commit are related to adding
support for this feature.
	modified:   matplot2.py