Updated `matplot2.py` to add option for equal axis in scatter plot and remove redundant axis setting in _setup_subplot_.
Detailed explanation:

In the updated version of `matplot2.py`, I added an optional argument `orthonormal` to the `scatter()` function. If this argument is set to
True, the script will set the axes to be equal using `plt.axis("equal")`. I believe this addition will provide more flexibility for users and
make the code more versatile.

Additionally, in the `_setup_subplot_()` function, there was a redundant axis setting for the case when `orthonormal` is True. This has been
removed to avoid confusion and potential inconsistencies. I believe these changes will improve the overall functionality of the script.
	modified:   matplot2.py